Removed unnecessary work from lazyFileNamesAndServices
Moved it to `createCoreMutationsProvider`, since this is the only time it's actually used.

Fixes #97.
